Revert "Fixes and updates"
This reverts commit b739af043b71f42ea16cf01e142f30ef9738980f.

+++ Ghidra/Features/Base/src/main/java/ghidra/app/util/viewer/util/MemoryBlockMap.java
@@ -44,7 +44,7 @@ public class MemoryBlockMap implements AddressPixelMap {
 		for (MemoryBlock block : blocks) {
 			totalSize += block.getSize();
 		}
-		addressesPerPixel = (float) totalSize / width;
+		addressesPerPixel = (float) totalSize / (float) width;
 		for (int i = 0; i < blocks.length; i++) {
 			pixels[i] = Math.round(blocks[i].getSize() / addressesPerPixel);
 		}
