+++ Ghidra/Framework/SoftwareModeling/src/main/java/ghidra/program/database/references/RefListV0.java
@@ -339,7 +339,7 @@ class RefListV0 extends RefList {
 	}
 
 	private byte findHighestRefLevel(byte currentRefLevel) {
-		byte maxLevel = -1;
+		byte maxLevel = (byte) -1;
 		ReferenceIterator it = getRefs();
 		while (it.hasNext()) {
 			Reference ref = it.next();
